Adding support to get test information via the api
I am trying to use in-app unit tests, and to validate whether or not
the test passed in-line.

To do this, I added a getCaseStatus endpoint.  This will return the
status for the case, agent tuple (if it has been run), otherwise it
will wait for it to be run and then call it once its successful.

The other added endpoint, getCaseInfo will map the case number to the
caseId as well as returning a brief description.

For the brief description, we call caseClassToPrettyDescription which
truncates whitespace, and returns everything to the left of the first
'<' since either invalid unicode or HTML follows.